change int to uint in test

+++ base62_test.go
@@ -1,29 +1,30 @@
 package base62
 
 import (
-  "fmt"
-  "testing"
-  assert "github.com/pilu/miniassert"
+	"fmt"
+	"testing"
+
+	assert "github.com/pilu/miniassert"
 )
 
 func TestEncode(t *testing.T) {
-  assert.Equal(t, "0", Encode(0))
-  assert.Equal(t, "1B", Encode(99))
+	assert.Equal(t, "0", Encode(0))
+	assert.Equal(t, "1B", Encode(99))
 }
 
 func TestDecode(t *testing.T) {
-  assert.Equal(t, 0,  Decode("0"))
-  assert.Equal(t, 99, Decode("1B"))
+	assert.Equal(t, uint(0), Decode("0"))
+	assert.Equal(t, uint(99), Decode("1B"))
 }
 
 func ExampleEncode() {
-  fmt.Println(Encode(99))
-  // Output:
-  // 1B
+	fmt.Println(Encode(99))
+	// Output:
+	// 1B
 }
 
 func ExampleDecode() {
-  fmt.Println(Decode("1B"))
-  // Output:
-  // 99
+	fmt.Println(Decode("1B"))
+	// Output:
+	// 99
 }
